+++ packages/app/src/lib/sanity.tsx
@@ -103,6 +103,7 @@ export function getImageProps<T extends ImageBlock>(
     alt,
     width,
     height,
+    extension: asset.extension,
   };
 }
